requirements.txt
alembic==1.8.1
autopep8==1.7.0
click==8.1.3
colorama==0.4.5
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.1
marshmallow-sqlalchemy==0.28.1
packaging==21.3
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyparsing==3.0.9
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.40
toml==0.10.2
Werkzeug==2.2.2
zipp==3.8.1
